+++ web_server/WSP-Pages/index.html
@@ -20,7 +20,6 @@
     <!-- Core theme CSS (includes Bootstrap)-->
     <link href="css/styles.css" rel="stylesheet" />
     <link href="css/index.css" rel="stylesheet" />
-    <link href="css/index-style.css" rel="stylesheet" />
 </head>
 
 <body id="page-top">
@@ -37,22 +36,30 @@
                     <li class="nav-item"><a class="nav-link" href="document.html">Document</a></li>
                     <li class="nav-item"><a class="nav-link" href="#contact">Contact</a></li>
                     <li class="nav-item">
-                        <a class="btn btn-outline-success d-xl-none d-lg-none mx-0 mb-3 mt-2" id="signUp" data-bs-toggle="modal" data-bs-target="#modalSignUp">
+                        <a class="btn btn-outline-success mx-0 mb-3 mt-2 d-lg-none" id="signUp" data-bs-toggle="modal" data-bs-target="#modalSignUp">
                             Sign Up
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="btn btn-success d-xl-none d-lg-none mx-0 mb-3" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
+                        <a class="btn btn-success mx-0 d-lg-none" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
                             Sign In
+                        </a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="btn btn-danger mx-0 mb-3 mt-2 d-lg-none d-none" id="logOut">
+                            Log out
                         </a>
                     </li>
                 </ul>
             </div>
-            <a class="btn btn-outline-success d-none d-lg-block d-xl-block" id="signUp" data-bs-toggle="modal" data-bs-target="#modalSignUp">
+            <a class="nav-base btn btn-outline-success d-none d-lg-block" id="signUp" data-bs-toggle="modal" data-bs-target="#modalSignUp">
                 Sign Up
             </a>
-            <a class="btn btn-success d-none d-lg-block d-xl-block" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
+            <a class="nav-base btn btn-success d-none d-lg-block" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
                 Sign In
+            </a>
+            <a class="nav-base btn btn-danger d-none" id="logOut">
+                Log out
             </a>
         </div>
     </nav>
@@ -77,7 +84,7 @@
         </div>
     </header>
     <!-- Contact-->
-    <section class="page-section" id="contact">
+    <section class="page-section pb-1" id="contact">
         <div class="container px-4 px-lg-5">
             <!--Google map-->
             <div id="row map-container-google" class="z-depth-1-half mb-5 map-container">
@@ -89,13 +96,13 @@
                 <div class="col-lg-8 col-xl-6 text-center">
                     <h2 class="mt-0">Let's Get In Touch!</h2>
                     <hr class="divider" />
-                    <p class="text-muted mb-5">
+                    <p class="text-muted mb-3">
                         Your feedback is valuable to us, as we work to improve both our products and our services to
                         you.
                     </p>
                 </div>
             </div>
-            <div class="row gx-4 gx-lg-5 justify-content-center mb-5">
+            <div class="row gx-4 gx-lg-5 justify-content-center">
                 <div class="col-lg-6">
                     <!-- * * * * * * * * * * * * * * *-->
                     <!-- * * SB Forms Contact Form * *-->
@@ -104,7 +111,7 @@
                     <!-- To make this form functional, sign up at-->
                     <!-- https://startbootstrap.com/solution/contact-forms-->
                     <!-- to get an API token!-->
-                    <form id="contactForm" action="./php/message.php">
+                    <form class="d-none" id="contactForm" action="./php/message.php">
                         <div class="align-items-center justify-content-center text-center">
                             <div class="form-floating mb-3">
                                 <select class="form-control" name="type">
@@ -119,7 +126,7 @@
                             <div class="form-floating mb-3">
                                 <textarea class="form-control" id="message" type="text" name="message"
                                     placeholder="Enter your message here..." style="height: 10rem" required></textarea>
-                                <label for="message">Message</label>
+                                <label for="message">Description of your idea or suggestion.</label>
                             </div>
                             <button class="btn btn-primary btn-xl" id="submitBtn" type="submit">
                                 Submit
@@ -128,12 +135,20 @@
                     </form>
                 </div>
             </div>
-            <!-- <div class="row gx-4 gx-lg-5 justify-content-center">
-                <div class="col-lg-4 text-center mb-5 mb-lg-0">
-                    <i class="bi-phone fs-2 mb-3 text-muted"></i>
-                    <div>+1 (555) 123-4567</div>
+            <div id="feedbackSignIn" class="row gx-4 gx-lg-5 h-100 align-items-center justify-content-center text-center">
+                <div class="col-lg-8 align-self-baseline">
+                    <a class="btn btn-success btn-xl" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
+                        Sign In
+                    </a>
                 </div>
-            </div> -->
+                <label class="text-muted mt-1">If you want to feedback sign In first.</label>
+            </div>
+            <div class="row gx-4 gx-lg-5 justify-content-center">
+                <div class="col-lg-4 text-center my-3 mb-lg-0">
+                    <i class="bi-phone fs-2 mb-3 text-muted"></i>
+                    <div>+1 (555) 33-45678</div>
+                </div>
+            </div>
         </div>
     </section>
     <!-- Sign Up Modal -->
@@ -162,12 +177,6 @@
                                 <input type="password" class="form-control" name="password" placeholder="Required"
                                     autocomplete="off" required />
                             </div>
-                            <!-- <div class="col-lg-6 form-outline mb-2">
-                                <label class="form-label mb-0" for="PatientName">Patient Name</label>
-                                <input type="text" class="form-control" name="patientName"
-                                    placeholder="Patient Name" autocomplete="off" />
-                                <label class="d-inline form-label" style="font-size: 0.8rem;">Non-essential.</label>
-                            </div> -->
                             <div class="col-lg-12 form-outline mb-2">
                                 <label class="form-label mb-0" for="emailAddress">Email Address </label>
                                 <input type="email" class="form-control" name="emailAddress" placeholder="Email Address"
@@ -252,13 +261,13 @@
     <!-- SimpleLightbox plugin JS-->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/SimpleLightbox/2.1.0/simpleLightbox.min.js"></script>
     <!-- Core theme JS-->
-    <script src="js/scripts.js"></script>
+    <script src="js/index-base.js"></script>
     <!-- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *-->
     <!-- * *                               SB Forms JS                               * *-->
     <!-- * * Activate your form at https://startbootstrap.com/solution/contact-forms * *-->
     <!-- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *-->
     <!-- Self page JS-->
-    <script src="./js/index.js"></script>
+    <script type="module" src="./js/index-script.js"></script>
 </body>
 
 </html>
