+++ web_server/WSP-Pages/js/index-base.js
