+++ web_server/WSP-Pages/paint/js/paint.js
@@ -530,4 +530,19 @@ class Paint {
     }
 }
 
+// Check image type
+function checkFiletype(file) {
+    if (file.type.match('image/jpg|image/jpeg|image/png')) {
+        return false;
+    }
+    return true;
+}
+// Check image size
+function checkFilesize(file) {
+    if (file.size < 5242880) {
+        return false;
+    }
+    return true;
+}
+
 export { Paint };
