+++ web_server/WSP-Pages/paint/js/style.js
@@ -17,11 +17,11 @@ function areatextPosition(string = null) {
 function loginStatus(status) {
     if (status) {
         document.querySelector('#signUp').classList.add("d-none");
-        document.querySelector('#signIN').classList.add("d-none");
+        document.querySelector('#signIn').classList.add("d-none");
         document.querySelector('#logOut').classList.remove("d-none");
     } else {
         document.querySelector('#signUp').classList.remove("d-none");
-        document.querySelector('#signIN').classList.remove("d-none");
+        document.querySelector('#signIn').classList.remove("d-none");
         document.querySelector('#logOut').classList.add("d-none");
     }
 }
