+++ web_server/WSP-Pages/paint/main.html
@@ -80,7 +80,7 @@
                             </a>
                         </li>
                         <li>
-                            <a class="btn btn-success" id="signIN" data-bs-toggle="modal" data-bs-target="#modalSignIn">
+                            <a class="btn btn-success" id="signIn" data-bs-toggle="modal" data-bs-target="#modalSignIn">
                                 Sign In
                             </a>
                         </li>
@@ -94,7 +94,7 @@
                 <div class="btn-group" role="group">
                     <span class="d-inline-block float-end" tabindex="0" data-bs-toggle="popover"
                         data-bs-trigger="focus hover" data-bs-html="true"
-                        data-bs-content="If the scale is set, then you can start automatic or manual calculation area.E">
+                        data-bs-content="If the scale is set, then you can start automatic or manual calculation area.">
                         <button type="button" class="btn btn-outline-secondary" id="rulerBtn">
                             <i class="bi bi-rulers"></i>
                         </button>
@@ -102,16 +102,8 @@
                 </div>
             </div>
             <div>
-                <!-- <div class="btn-group" role="group">
-                    <button type="button" class="btn btn-outline-secondary" id="testBtn">
-                        1</i>
-                    </button>
-                    <button type="button" class="btn btn-outline-secondary" id="test2Btn">
-                        2</i>
-                    </button>
-                </div> -->
                 <div class="btn-group" role="group">
-                    <button type="button" class="btn btn-outline-success" id="chartBtn">
+                    <button type="button" class="btn btn-outline-success d-none" id="chartBtn">
                         <i class="bi bi-bar-chart-fill"></i>
                     </button>
                 </div>
@@ -414,7 +406,7 @@
     <script type="module" src="./js/paint-jquery.js"></script>
     <!-- Self JS -->
     <script type="module" src="./js/paint-script.js"></script>
-    <script type="module" src="../js/login.js"></script>
+    <script type="module" src="./js/paint-login.js"></script>
 </body>
 
 </html>
